style: lighten theme and add scroll animation
